Fix control plane migration of ContainerRuntimes
