+++ pkg/controller/actuator.go
@@ -5,8 +5,12 @@
 package controller
 
 import (
+	"context"
+	"time"
+
 	extensionscontroller "github.com/gardener/gardener/extensions/pkg/controller"
 	"github.com/gardener/gardener/extensions/pkg/controller/containerruntime"
+	"github.com/gardener/gardener/pkg/utils/managedresources"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -22,4 +26,19 @@ func NewActuator(c client.Client, chartRendererFactory extensionscontroller.Char
 		chartRendererFactory: chartRendererFactory,
 		client:               c,
 	}
+}
+
+func (a *actuator) deleteManagedResource(ctx context.Context, namespace, managedResourceName string, forceDelete bool) error {
+	if err := managedresources.Delete(ctx, a.client, namespace, managedResourceName, true); err != nil {
+		return err
+	}
+
+	if !forceDelete {
+		timeoutCtx, cancel := context.WithTimeout(ctx, 2*time.Minute)
+		defer cancel()
+
+		return managedresources.WaitUntilDeleted(timeoutCtx, a.client, namespace, managedResourceName)
+	}
+
+	return nil
 }
